+++ src/wrapper/clap/context.rs
@@ -1,14 +1,20 @@
 use atomic_refcell::AtomicRefMut;
+use clap_sys::ext::draft::remote_controls::{
+    clap_remote_controls_page, CLAP_REMOTE_CONTROLS_COUNT,
+};
+use clap_sys::id::{clap_id, CLAP_INVALID_ID};
+use clap_sys::string_sizes::CLAP_NAME_SIZE;
 use std::cell::Cell;
-use std::collections::VecDeque;
+use std::collections::{HashMap, VecDeque};
 use std::sync::Arc;
 
 use super::wrapper::{OutputParamEvent, Task, Wrapper};
 use crate::event_loop::EventLoop;
 use crate::prelude::{
     ClapPlugin, GuiContext, InitContext, ParamPtr, PluginApi, PluginNoteEvent, ProcessContext,
-    Transport,
+    RemoteControlsContext, RemoteControlsPage, RemoteControlsSection, Transport,
 };
+use crate::wrapper::util::strlcpy;
 
 /// An [`InitContext`] implementation for the wrapper.
 ///
@@ -47,6 +53,16 @@ pub(crate) struct WrapperGuiContext<P: ClapPlugin> {
     #[cfg(debug_assertions)]
     pub(super) param_gesture_checker:
         atomic_refcell::AtomicRefCell<crate::wrapper::util::context_checks::ParamGestureChecker>,
+}
+
+/// A [`RemoteControlsContext`] implementation for the wrapper. This is used during initialization
+/// to allow the plugin to declare remote control pages. This struct defines the pages in the
+/// correct format.
+pub(crate) struct RemoteControlPages<'a> {
+    param_ptr_to_hash: &'a HashMap<ParamPtr, u32>,
+    /// The remote control pages, as defined by the plugin. These don't reference any heap data so
+    /// we can store them directly.
+    pages: &'a mut Vec<clap_remote_controls_page>,
 }
 
 impl<P: ClapPlugin> Drop for WrapperInitContext<'_, P> {
@@ -223,5 +239,140 @@ impl<P: ClapPlugin> GuiContext for WrapperGuiContext<P> {
 
     fn set_state(&self, state: crate::wrapper::state::PluginState) {
         self.wrapper.set_state_object_from_gui(state)
+    }
+}
+
+/// A remote control section. The plugin can fill this with information for one or more pages.
+pub(crate) struct Section {
+    pages: Vec<Page>,
+}
+
+/// A remote control page. These are automatically split into multiple pages if the number of
+/// controls exceeds 8.
+pub(crate) struct Page {
+    name: String,
+    params: Vec<Option<ParamPtr>>,
+}
+
+impl<'a> RemoteControlPages<'a> {
+    /// Allow the plugin to define remote control pages and add them to `pages`. This does not clear
+    /// `pages` first.
+    pub fn define_remote_control_pages<P: ClapPlugin>(
+        plugin: &P,
+        pages: &'a mut Vec<clap_remote_controls_page>,
+        param_ptr_to_hash: &'a HashMap<ParamPtr, u32>,
+    ) {
+        // The magic happens in the `add_section()` function defined below
+        plugin.remote_controls(&mut Self {
+            pages,
+            param_ptr_to_hash,
+        });
+    }
+
+    /// Perform the boilerplate needed for creating and adding a new [`clap_remote_controls_page`].
+    /// If `params` contains more than eight parameters then any further parameters will be lost.
+    fn add_clap_page(
+        &mut self,
+        section: &str,
+        page_name: &str,
+        params: impl IntoIterator<Item = Option<ParamPtr>>,
+    ) {
+        let mut page = clap_remote_controls_page {
+            section_name: [0; CLAP_NAME_SIZE],
+            // Pages are numbered sequentially
+            page_id: self.pages.len() as clap_id,
+            page_name: [0; CLAP_NAME_SIZE],
+            param_ids: [CLAP_INVALID_ID; CLAP_REMOTE_CONTROLS_COUNT],
+            is_for_preset: false,
+        };
+        strlcpy(&mut page.section_name, section);
+        strlcpy(&mut page.page_name, page_name);
+
+        let mut params = params.into_iter();
+        for (param_id, param_ptr) in page.param_ids.iter_mut().zip(&mut params) {
+            // `param_id` already has the correct value if `param_ptr` is empty/a spacer
+            if let Some(param_ptr) = param_ptr {
+                *param_id = self.param_ptr_to_id(param_ptr);
+            }
+        }
+
+        nih_debug_assert!(
+            params.next().is_none(),
+            "More than eight parameters were passed to 'RemoteControlPages::add_page()', this is \
+             a NIH-plug bug."
+        );
+
+        self.pages.push(page);
+    }
+
+    /// Transform a `ParamPtr` to the associated CLAP parameter ID/hash. Returns -1/invalid
+    /// parameter and triggers a debug assertion when the parameter is not known.
+    fn param_ptr_to_id(&self, ptr: ParamPtr) -> clap_id {
+        match self.param_ptr_to_hash.get(&ptr) {
+            Some(id) => *id,
+            None => {
+                nih_debug_assert_failure!(
+                    "An unknown parameter was added to a remote control page, ignoring..."
+                );
+
+                CLAP_INVALID_ID
+            }
+        }
+    }
+}
+
+impl RemoteControlsContext for RemoteControlPages<'_> {
+    type Section = Section;
+
+    fn add_section(&mut self, name: impl Into<String>, f: impl FnOnce(&mut Self::Section)) {
+        let section_name = name.into();
+        let mut section = Section {
+            pages: Vec::with_capacity(1),
+        };
+        f(&mut section);
+
+        // The pages in the section may need to be split up into multiple pages if it defines more
+        // than eight parameters. This keeps the interface flexible for potential future expansion
+        // and makes manual paging unnecessary in some situations.
+        for page in section.pages {
+            if page.params.len() > CLAP_REMOTE_CONTROLS_COUNT {
+                for (subpage_idx, subpage_params) in
+                    page.params.chunks(CLAP_REMOTE_CONTROLS_COUNT).enumerate()
+                {
+                    let subpage_name = format!("{} {}", page.name, subpage_idx + 1);
+                    self.add_clap_page(
+                        &section_name,
+                        &subpage_name,
+                        subpage_params.iter().copied(),
+                    );
+                }
+            } else {
+                self.add_clap_page(&section_name, &page.name, page.params);
+            }
+        }
+    }
+}
+
+impl RemoteControlsSection for Section {
+    type Page = Page;
+
+    fn add_page(&mut self, name: impl Into<String>, f: impl FnOnce(&mut Self::Page)) {
+        let mut page = Page {
+            name: name.into(),
+            params: Vec::with_capacity(CLAP_REMOTE_CONTROLS_COUNT),
+        };
+        f(&mut page);
+
+        self.pages.push(page);
+    }
+}
+
+impl RemoteControlsPage for Page {
+    fn add_param(&mut self, param: &impl crate::prelude::Param) {
+        self.params.push(Some(param.as_ptr()));
+    }
+
+    fn add_spacer(&mut self) {
+        self.params.push(None);
     }
 }